+++ src/components/app/app.tsx
@@ -1,11 +1,11 @@
 import {BrowserRouter, Route, Routes} from 'react-router-dom';
 
-import MainPage from '../../pages/main/main-page';
+import MainPage from '../../pages/main-page/main-page';
 import {AppRoute} from '../../consts';
 import NotFoundPage from '../../pages/not-found/not-found';
 import LoginPage from '../../pages/login-page/login-page';
 import FavoritesPage from '../../pages/favorites-page/favorites-page';
-import OfferPage from '../../pages/offer/offer-page';
+import OfferPage from '../../pages/offer-page/offer-page';
 import PrivateRoute from '../private-route/private-route';
 import {useSelector} from 'react-redux';
 import {RootState} from '../../store';
